Move tip finding algorithms to their own files and add a general one
