+++ src/algorithms/distance_to_head.hpp
@@ -14,8 +14,6 @@ namespace algorithms {
 
 using namespace std;
 
-/// Find all of the nodes with no edges on their left sides.
-vector<handle_t> head_nodes(const HandleGraph* g);
 int32_t distance_to_head(handle_t h, int32_t limit, const HandleGraph* graph);
 /// Get the distance in bases from start of node to start of closest head node of graph, or -1 if that distance exceeds the limit.
 /// dist increases by the number of bases of each previous node until you reach the head node
