+++ src/algorithms/distance_to_tail.hpp
@@ -14,8 +14,6 @@ namespace algorithms {
 
 using namespace std;
 
-/// Find all of the nodes with no edges on their left sides.
-vector<handle_t> tail_nodes(const HandleGraph* g);
 int32_t distance_to_tail(handle_t h, int32_t limit, const HandleGraph* graph);
 /// Get the distance in bases from end of node to end of closest tail node of graph, or -1 if that distance exceeds the limit.
 /// dist increases by the number of bases of each previous node until you reach the head node
