+++ src/algorithms/topological_sort.hpp
@@ -22,12 +22,6 @@ namespace algorithms {
 
 using namespace std;
 
-/// Find all of the nodes with no edges on their left sides.
-vector<handle_t> head_nodes(const HandleGraph* g);
-
-/// Find all of the nodes with no edges on their right sides.
-vector<handle_t> tail_nodes(const HandleGraph* g);
-
 /**
  * Order and orient the nodes in the graph using a topological sort. The sort is
  * guaranteed to be machine-independent given the initial graph's node and edge
